Rearrange things nicely.
